requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
pg8000==1.31.2
python-docx==1.1.2
PyMuPDF==1.25.3
nltk==3.9.1
python-dotenv==1.0.1
Werkzeug==3.1.3
SQLAlchemy==2.0.36